We can make xmi-files now & "random" is added to model
New package dk.itu.dsl.roboprose.compiler can be run with sbt "run
[filepath]" and then it spits out an xmi-file in the same place with the
same name in the same place as the source.

Abstract class "randomizable" with attribute isRandom, and enum RANDOM is
added to ecore model and actions move and turn are now randomizable.
This is also added to the grammar.

A couple of example .roboprose and their .xmi files are also added.
"dancing-robot", pretty simple, and "crazy-robot" which is more complex.