Added "Endings" and restructured the model a bit
Removed a couple of Enums, because i learned some smart stuff about
xtext (the ?= operator) and unlearned some unsmart stuff.

Random is now a Boolean isRandom again.

"Endings" is a new class that is like a controller option that can be
set after a block of actions.
Three types of Ending are added:
"repeat" - Which should repeat the current block
"start over" - Which should return the robot to it's default behaviour
(main block)
"wait" - Which is like indefinite stop.

MOVE_DIRECTION and TURN_DIRECTION are kept as ENums, because we might
want to add more directions further down the line.